requirements.txt
Flask
spacy
scikit-learn
pdfplumber
python-docx
gunicorn